+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Dependencies: Updated to `async-kinesis-2.0.0`. Thanks, @hampsterx.
 - CDC: Added canonical SQL example for PostgreSQL from Ibis
 - CDC: Enabled loading DMS events from Kinesis streams and stream-dump files
+- CDC: Added subcommand `ctk dms table-mappings`
 
 ## 2025/05/13 v0.0.35
 - Added lost `pytest` dependencies to `cratedb-toolkit[testing]`